+++ client/src/modules/list/components/Items/ListItem/ListItem.scss
